# apiserver/model_generator.go
/*
 * Device Simulator app API
 *
 * API to access and configure the Device Simulator app
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apiserver

type Generator struct {
	Id int32 `json:"id,omitempty"`

	AssetId int32 `json:"asset_id"`

	Attribute string `json:"attribute"`

	FunctionType string `json:"function_type"`

	MinValue float64 `json:"min_value"`

	MaxValue float64 `json:"max_value"`

	IntervalSeconds int32 `json:"interval_seconds"`

	Frequency float64 `json:"frequency"`
}

// AssertGeneratorRequired checks if the required fields are not zero-ed
func AssertGeneratorRequired(obj Generator) error {
	elements := map[string]interface{}{
		"asset_id":         obj.AssetId,
		"attribute":        obj.Attribute,
		"function_type":    obj.FunctionType,
		"min_value":        obj.MinValue,
		"max_value":        obj.MaxValue,
		"interval_seconds": obj.IntervalSeconds,
		"frequency":        obj.Frequency,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertGeneratorConstraints checks if the values respects the defined constraints
func AssertGeneratorConstraints(obj Generator) error {
	return nil
}